actions: centralize in sharedsec
