+++ internal/appsec/emitter/sharedsec/actions_test.go
@@ -235,3 +235,81 @@ func TestNewRedirectRequestAction(t *testing.T) {
 		require.Equal(t, "Redirected", string(body))
 	})
 }
+
+func TestNewBlockParams(t *testing.T) {
+	for name, tc := range map[string]struct {
+		params   map[string]any
+		expected blockActionParams
+	}{
+		"block-1": {
+			params: map[string]any{
+				"status_code": "403",
+				"type":        "auto",
+			},
+			expected: blockActionParams{
+				Type:       "auto",
+				StatusCode: 403,
+			},
+		},
+		"block-2": {
+			params: map[string]any{
+				"status_code": "405",
+				"type":        "html",
+			},
+			expected: blockActionParams{
+				Type:       "html",
+				StatusCode: 405,
+			},
+		},
+	} {
+		t.Run(name, func(t *testing.T) {
+			actionParams, err := blockParamsFromMap(tc.params)
+			require.NoError(t, err)
+			require.Equal(t, tc.expected.Type, actionParams.Type)
+			require.Equal(t, tc.expected.StatusCode, actionParams.StatusCode)
+		})
+	}
+}
+
+func TestNewRedirectParams(t *testing.T) {
+	for name, tc := range map[string]struct {
+		params   map[string]any
+		expected redirectActionParams
+	}{
+		"redirect-1": {
+			params: map[string]any{
+				"status_code": "308",
+				"location":    "/redirected",
+			},
+			expected: redirectActionParams{
+				Location:   "/redirected",
+				StatusCode: 308,
+			},
+		},
+		"redirect-2": {
+			params: map[string]any{
+				"status_code": "303",
+				"location":    "/tmp",
+			},
+			expected: redirectActionParams{
+				Location:   "/tmp",
+				StatusCode: 303,
+			},
+		},
+		"no-location": {
+			params: map[string]any{
+				"status_code": "303",
+			},
+			expected: redirectActionParams{
+				Location:   "",
+				StatusCode: 303,
+			},
+		},
+	} {
+		t.Run(name, func(t *testing.T) {
+			actionParams, err := redirectParamsFromMap(tc.params)
+			require.NoError(t, err)
+			require.Equal(t, tc.expected, actionParams)
+		})
+	}
+}
